creando la clase CreateRestaurantUseCase

+++ src/main/java/com/pragma/app/domain/model/Restaurant.java
@@ -17,5 +17,5 @@ public class Restaurant{
     private String address;
     private String cellPhoneRestaurant;
     private String urlLogo;
-    private Long idOwner;
+    private int idOwner;
 }
